requirements.txt
annotated-types==0.7.0
anyio==4.12.1
astrapy==2.1.0
async-timeout==4.0.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
cryptography==46.0.5
deprecation==2.1.0
dnspython==2.7.0
exceptiongroup==1.3.1
filetype==1.2.0
google-ai-generativelanguage==0.10.0
google-api-core==2.30.0
google-auth==2.49.0
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.28
langchain-astradb==0.6.1
langchain-core==0.3.83
langchain-google-genai==2.1.12
langchain-text-splitters==0.3.11
langchainhub==0.1.21
langgraph==0.6.11
langgraph-checkpoint==2.1.2
langgraph-prebuilt==0.6.5
langgraph-sdk==0.2.9
langsmith==0.4.37
numpy==2.0.2
orjson==3.11.5
ormsgpack==1.11.0
packaging==24.2
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.16.0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
SQLAlchemy==2.0.48
tenacity==9.1.2
toml==0.10.2
types-requests==2.32.4.20260107
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid6==2025.0.1
uuid_utils==0.14.1
xxhash==3.6.0
zstandard==0.25.0
